requirements.txt
akshare
pandas
numpy
matplotlib
openpyxl
streamlit
plotly
tushare
psycopg2-binary
